Code cleanup

+++ core/src/main/java/org/javarosa/core/model/data/DecimalData.java
@@ -95,7 +95,7 @@ public class DecimalData implements IAnswerData {
 
     @Override
 	public UncastData uncast() {
-		return new UncastData(((Double)getValue()).toString());
+		return new UncastData(getValue().toString());
 	}
 
     @Override
